chore: add styled-components, format
